led: Move colour/time ownership to Patterns
This makes a large amount of the boilerplate in LightStrip disappear,
by making each Pattern own their colour and start time.

This also makes each Pattern take a HsvColour instead of arbitrary Hsvs.

A `clock` default field is added to allow dependency injection in tests
(but also a local lookup is faster than a global lookup).